gcloud/agent/requirements.txt
flask==3.0.0
google-cloud-storage==2.14.0
gunicorn==21.2.0
google-generativeai==0.8.3
requests==2.31.0
